codes/requirements.txt
tensorflow >= 1.11.0   # CPU Version of TensorFlow.
# tensorflow-gpu  >= 1.11.0  # GPU version of TensorFlow.
pytorch_pretrained_bert
torch
pyknp
bert-serving-server
bert-serving-client
flask